requirements.txt
http://github.com/django/django/tarball/stable/1.5.x#egg=Django
South
Pillow
django-tastypie
django_compressor
cssmin
jsmin
django-images
http://github.com/hcarvalhoalves/django-taggit/tarball/master#egg=django-taggit